Revise when-let "gating" logic
agent-shell uses a lot of the following pattern to logically gate certain blocks:

```elisp
(when ((_ (some-condition-p)))
  ...)
```

Logically, this is correct. However, on Emacs 29, this results in some byte-compilation errors
related to `_` actually not **technically** remaining unused under the hood after macro expansion.

As it turns out, the following semantics for `if-let` and friends have been valid since Emacs
26 ([citation](https://emacs.stackexchange.com/a/64913)):

```elisp
(when (((some-condition-p)))
  ...)
```

From `if-let*` docs:

```
An element can additionally be of the form (VALUEFORM), which is evaluated and checked for nil;
i.e. SYMBOL can be omitted if only the test result is of interest.  It can also be of the form
SYMBOL, then the binding of SYMBOL is checked for nil.
```

This PR updates all use of the "underscore assignment" pattern in agent-shell to take advantage of
the aforementioned semantics.